// PricePresenter - cast result to avoid string on 0 values

+++ src/Adapter/Product/PricePresenter.php
@@ -8,7 +8,7 @@ class PricePresenter
 {
     public function convertAmount($price)
     {
-        return Tools::convertPrice($price);
+        return (float)Tools::convertPrice($price);
     }
 
     public function format($price)
